Added Identity object tests
Added the latest version of the OpenADK again and upgraded the Solution
file to fix a deployment issue. Added test Iterator, Publisher and
Subscriber for the Identity object. Updated the Agent configuration
files to use a SIF 2.6 compliant Zone on the NSIP ZIS (required for the
Identity object which is AU 2.4+).